requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
pinecone-client==3.0.2
yt-dlp==2024.3.10
requests==2.31.0
